requirements.txt
cffi==1.15.0
click==8.0.1
colorama==0.4.4
cryptography==36.0.0
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.3.1
MarkupSafe==2.0.1
Pillow==9.0.1
psycopg2==2.9.1
pycparser==2.21
PyJWT==2.3.0
six==1.16.0
Werkzeug==2.0.1
